ticket_app/requirements.txt
django
djangorestframework
gunicorn
drf_yasg